feat: update InputNode to use int for Value property, remove unused deserialization event in FlowCanvas, and refactor FlowGraph serialization to use System.Text.Json

+++ examples/FlowStateBlazorServer/Components/Nodes/InputNode.razor.cs
@@ -10,7 +10,7 @@ public partial class InputNode
 {
 
     [Parameter]
-    public long Value { get; set; }
+    public int Value { get; set; }
 
     public override ValueTask ExecuteAsync(FlowExecutionContext context)
     {
